requirements.txt
python-telegram-bot
requests
deemix
